// turbo/generators/tokens/acf.type.ts
// generated source
export type AcfTokenFields = {
	key: string
	title: string
	fields: Array<{
		key: string
		label: string
		name: string
		'aria-label': string
		type: string
		instructions: string
		required: number
		conditional_logic: number
		wrapper: {
			width: string
			class: string
			id: string
		}
		choices: {
			'1'?: string
			'2'?: string
			'3'?: string
			'4'?: string
			'5'?: string
			'6'?: string
			'7'?: string
			'8'?: string
			'9'?: string
			body?: string
			heading?: string
			brand?: string
			primary?: string
			secondary?: string
			accent?: string
			px?: string
			'0'?: string
			'0.5'?: string
			'1.5'?: string
			'2.5'?: string
			'3.5'?: string
			'10'?: string
			'11'?: string
			'12'?: string
			'14'?: string
			'16'?: string
			'20'?: string
			'24'?: string
			'28'?: string
			'32'?: string
			'36'?: string
			'40'?: string
			'44'?: string
			'48'?: string
			'52'?: string
			'56'?: string
			'60'?: string
			'64'?: string
			'72'?: string
			'80'?: string
			'96'?: string
			'0.25'?: string
			'0.75'?: string
			'1.25'?: string
		}
		default_value: any
		return_format: string
		multiple: number
		allow_null: number
		allow_in_bindings: number
		ui: number
		show_in_graphql: number
		graphql_description: string
		graphql_field_name: string
		graphql_non_null: number
		ajax: number
		placeholder: string
	}>
	location: Array<
		Array<{
			param: string
			operator: string
			value: string
		}>
	>
	menu_order: number
	position: string
	style: string
	label_placement: string
	instruction_placement: string
	hide_on_screen: string
	active: boolean
	description: string
	show_in_rest: number
	show_in_graphql: number
	graphql_field_name: string
	map_graphql_types_from_location_rules: number
	graphql_types: string
	modified: number
}